Kiran

+++ Modules/Cart/app/Http/Controllers/CartController.php
@@ -52,19 +52,19 @@ class CartController extends Controller
         $product = Product::find($request->product_id);
 
         // Check if the product is in stock
-        if ($product->stock <= 0) {
-            return redirect()->route('cart.index')->with('error', 'Product is out of stock');
-        }
+        // if ($product->stock <= 0) {
+        //     return redirect()->route('cart.index')->with('error', 'Product is out of stock');
+        // }
 
         // Decrease the stock of the product
-        $product->stock--;
-        $product->save();
+        // $product->stock--;
+        // $product->save();
 
         $item = new Cart();
         $item->user_id = auth()->user()->id;
         $item->product_id = $request->product_id;
         $item->save();
-        return redirect()->route('cart.index')->with('success', 'Product added to cart successfully');
+        return back()->with('success', 'Product added to cart successfully');
     }
 
     /**
@@ -103,11 +103,16 @@ class CartController extends Controller
         $item->save();
         return back()->with('success', 'Product added to wishlist successfully');
     }
-    public function removeFromWishlist($id)
+    public function removeFromWishlist($productId)
     {
-        $wishlist = Wishlist::find($id);
-        $wishlist->delete();
-        return back()->with('success', 'Product removed from wishlist successfully');
+        $userId = auth()->user()->id;
+        $wishlist = Wishlist::where('user_id', $userId)->where('product_id', $productId)->first();
+        if ($wishlist) {
+            $wishlist->delete();
+            return back()->with('success', 'Product removed from wishlist successfully');
+        } else {
+            return back()->with('error', 'Product not found in wishlist');
+        }
     }
     public function order()
     {
